only keep needed functionality of ConversionFinder
